+++ src/main/java/nl/itslars/kosmos/World.java
@@ -3,6 +3,7 @@ package nl.itslars.kosmos;
 import com.google.common.io.Files;
 import lombok.Getter;
 import nl.itslars.kosmos.enums.Dimension;
+import nl.itslars.kosmos.leveldb.LevelDB;
 import nl.itslars.kosmos.objects.entity.Player;
 import nl.itslars.kosmos.objects.settings.LevelDatFile;
 import nl.itslars.kosmos.objects.world.ChunkPreset;
